+++ screens/home/parallax-tablet/index.js
@@ -268,6 +268,7 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                         src={`${assetDomain}${
                             sectionOne?.familyImageMobile?.url ?? ''
                         }`}
+                        alt="family-image"
                     />
                     <div
                         id="icon-inpatient"
@@ -284,12 +285,10 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                             }}
                         >
                             <img
-                                // style={{
-                                //     width: `${iconHeight.inpatient}px`,
-                                // }}
                                 src={`${assetDomain}${
                                     sectionOne?.inpatientIcon?.url ?? ''
                                 }`}
+                                alt="inpatient-icon"
                             />
                         </div>
                     </div>
@@ -308,12 +307,10 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                             }}
                         >
                             <img
-                                // style={{
-                                //     width: `${iconHeight.dental}px`,
-                                // }}
                                 src={`${assetDomain}${
                                     sectionOne?.dentalIcon?.url ?? ''
                                 }`}
+                                alt="dental-icon"
                             />
                         </div>
                     </div>
@@ -332,12 +329,10 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                             }}
                         >
                             <img
-                                // style={{
-                                //     width: `${iconHeight.maternity}px`,
-                                // }}
                                 src={`${assetDomain}${
                                     sectionOne?.maternityIcon?.url ?? ''
                                 }`}
+                                alt="maternity-icon"
                             />
                         </div>
                     </div>
@@ -359,6 +354,7 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                                 src={`${assetDomain}${
                                     sectionOne?.outpatientIcon?.url ?? ''
                                 }`}
+                                alt="outpatient-icon"
                             />
                         </div>
                     </div>
@@ -386,6 +382,7 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                         src={`${assetDomain}${
                             sectionTwo?.backgroundImageMobile?.url ?? ''
                         }`}
+                        alt="living-room-image"
                     />
 
                     <img
@@ -394,6 +391,7 @@ export default function ParallaxTablet({ sectionOne, sectionTwo }) {
                         src={`${assetDomain}${
                             sectionTwo?.familyImage?.url ?? ''
                         }`}
+                        alt="family-revolve-image"
                     />
                 </div>
             )}