+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 > 2021-11-16
 
 -   Read SEO title, description, and keywords from strapi for home, product, and download page
+-   Add home page alt on image and set min height for mobile home
 
 ### [v1.0.1]